requirements.txt
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
exceptiongroup==1.3.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.2
ibm-cos-sdk-core==2.14.2
ibm-cos-sdk-s3transfer==2.14.2
ibm-watsonx-ai==1.3.26
idna==3.10
jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lomond==0.3.3
markupsafe==3.0.2
narwhals==1.44.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pymupdf==1.26.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.46.0
streamlit-cookies-manager==0.2.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0